[FIX] Compatibility of BE module for TYPO3 CMS 7.x
Make the backend module working with 6.2 and 7.x (7.5/7.6).

Instead of hard coding the urls for adding, editing, hiding and unhiding
nodes, get them via an ajax call. The target urls are generated for the
different versions in the tx_caretaker_Utility.

Adjust the helper class and add a default test configuration to avoid an
exception while rendering the new test form.

Adjust initialization and main javascript of node tree and node info
module to use the new services.

Adjust the TCA and use the new rendering configuration for select and
group fields.